src/test/java/com/leetcode/dp/medium/CountingBits_338Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class CountingBits_338Test {

  private final CountingBits_338 solution = new CountingBits_338();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(2, new int[] {0, 1, 1}), arguments(5, new int[] {0, 1, 1, 2, 1, 2}));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void countBits(int num, int[] expectedResult) {
    int[] actualResult = solution.countBits(num);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
